requirements.txt
PyQt6==6.7.0
pyqtgraph==0.13.3
pyserial==3.5
numpy==1.26.4
